# docs/swagger.yaml
basePath: /api
definitions:
...
  common.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  payroll.GeneratePayrollRequest:
    properties:
      allowances:
        type: number
      bonus:
        type: number
      deductions:
        type: number
      employee_id:
        type: string
      month:
        maximum: 12
        minimum: 1
        type: integer
      tax:
        type: number
      year:
        maximum: 2100
        minimum: 2000
        type: integer
    required:
    - employee_id
    type: object
host: localhost:8080
info:
  contact:
    email: ravi@example.com
    name: Raviraj Sahu
  description: HRMS Backend API using Go, Gin, GORM & PostgreSQL.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: HRMS API
  version: "1.0"
paths:
...
  /payroll:
    get:
      description: Get all payroll records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get All Payrolls
      tags:
      - Payroll
  /payroll/{id}:
    delete:
      description: Delete a payroll record
      parameters:
      - description: Payroll ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Payroll
      tags:
      - Payroll
    get:
      description: Get payroll details by payroll ID
      parameters:
      - description: Payroll ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Payroll By ID
      tags:
      - Payroll
  /payroll/{id}/pay:
    put:
      description: Mark a payroll as paid
      parameters:
      - description: Payroll ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Mark Payroll as Paid
      tags:
      - Payroll
  /payroll/employee/{employee_id}:
    get:
      description: Get payroll history of an employee
      parameters:
      - description: Employee ID
        in: path
        name: employee_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Payroll By Employee
      tags:
      - Payroll
  /payroll/generate:
    post:
      consumes:
      - application/json
      description: Generate payroll for an employee
      parameters:
      - description: Generate Payroll Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payroll.GeneratePayrollRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - BearerAuth: []
      summary: Generate Payroll
      tags:
      - Payroll
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"